booking_system/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
django-filter==24.2
django-sendgrid-v5==1.2.3
psycopg2-binary==2.9.9
gunicorn==22.0.0
zoomus==1.2.1
django[bcrypt]==5.0.6
python-dotenv==1.0.1